scraping/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.1.3
python-dotenv==1.0.0
supabase==2.0.2
schedule==1.2.0
selenium==4.15.2
webdriver-manager==4.0.1
